Set up the procedure on the delete method for active effect, use custom icon
