+++ module/documents/effects/active-effect.mjs
@@ -2,6 +2,7 @@ import { FUActor } from '../actors/actor.mjs';
 import { FUItem } from '../items/item.mjs';
 import { SYSTEM } from '../../helpers/config.mjs';
 import { ExpressionContext, Expressions } from '../../expressions/expressions.mjs';
+import { Effects } from './effects.mjs';
 
 const CRISIS_INTERACTION = 'CrisisInteraction';
 const EFFECT_TYPE = 'type';
@@ -105,6 +106,7 @@ const PRIORITY_CHANGES = [
  * @property {Boolean} isTemporary Describe whether the ActiveEffect has a temporary duration based on combat turns or rounds.
  * @property {Boolean} isEmbedded Test whether this Document is embedded within a parent Document
  * @property {String} uuid
+ * @property {EffectChangeData[]} changes - The array of EffectChangeData objects which the ActiveEffect applies
  * @remarks https://foundryvtt.com/api/classes/client.ActiveEffect.html
  * */
 export class FUActiveEffect extends ActiveEffect {
@@ -115,6 +117,27 @@ export class FUActiveEffect extends ActiveEffect {
 	async _preCreate(data, options, user) {
 		this.updateSource({ name: game.i18n.localize(data.name) });
 		return super._preCreate(data, options, user);
+	}
+
+	_onDelete(options, userId) {
+		const owner = options.parent;
+		if (owner instanceof FUActor) {
+			this.changes.forEach((change) => {
+				if (change.key === Effects.customAttributeKeys.linkedEffect) {
+					const path = change.value.split('.');
+					const itemId = path[0];
+					const item = owner.items.find((i) => i.id === itemId);
+
+					const effectId = path[1];
+					const linkedEffect = item.effects.get(effectId);
+					if (linkedEffect) {
+						console.debug(`Will remove the linked effect ${linkedEffect}`);
+						linkedEffect.delete();
+					}
+				}
+			});
+		}
+		super._onDelete(options, userId);
 	}
 
 	get isSuppressed() {
@@ -170,7 +193,6 @@ export class FUActiveEffect extends ActiveEffect {
 				// Second, evaluate with our custom expressions
 				const context = ExpressionContext.resolveTarget(target);
 				const value = Expressions.evaluate(expression, context);
-				console.debug('Substituting active effect change variable:', change.value, value);
 				change.value = String(value ?? 0);
 			} catch (e) {
 				console.error(e);
